+++ yaramail/__init__.py
@@ -8,14 +8,13 @@ import zipfile
 
 import yara
 import pdftotext
-from publicsuffix2 import get_sld
 
 from mailsuite.utils import parse_email, from_trusted_domain, decode_base64
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
 
-__version__ = "2.1.0"
+__version__ = "3.0.0"
 
 
 delimiters = ["\"", "'", "`", "\*\*",
@@ -145,7 +144,7 @@ class MailScanner(object):
                  attachment_rules: Union[str, IOBase, yara.Rules] = None,
                  passwords: Union[List[str], IOBase, str] = None,
                  max_zip_depth: int = None,
-                 yara_optional_domains: Union[List[str], IOBase, str] = None,
+                 yara_safe_optional_domains: Union[List[str], IOBase, str] = None,
                  allow_multiple_authentication_results: bool = False,
                  use_authentication_results_original: bool = False):
         """
@@ -163,9 +162,6 @@ class MailScanner(object):
             passwords: A list of passwords to use when attempting to scan \
             password-protected files
             max_zip_depth: Number of times to recurse into nested ZIP files
-            trusted_domains: A list of message From domains that return a \
-            ``safe`` verdict if the domain is authenticated and no YARA \
-            categories match other than ``safe``
             allow_multiple_authentication_results: Allow multiple \
             ``Authentication-Results-Original`` headers when checking \
             authentication results
@@ -226,7 +222,7 @@ class MailScanner(object):
         self.passwords += ["malware", "infected"]
         self.passwords = _deduplicate_list(self.passwords)
         self.max_zip_depth = max_zip_depth
-        self.trusted_domains = _input_to_str_list(yara_optional_domains)
+        self.yara_safe_optional_domains = _input_to_str_list(yara_safe_optional_domains)
         allow_multi_auth = allow_multiple_authentication_results
         self.allow_multiple_authentication_results = allow_multi_auth
         use_og_auth = use_authentication_results_original
@@ -266,7 +262,9 @@ class MailScanner(object):
         with zipfile.ZipFile(payload) as zip_file:
             for name in zip_file.namelist():
                 if passwords is None:
-                    passwords == []
+                    passwords = []
+                if None not in passwords:
+                    passwords = [None] + passwords
                 for password in passwords:
                     if isinstance(password, str):
                         password = password.encode("utf-8")
@@ -332,7 +330,7 @@ class MailScanner(object):
 
         if passwords is None:
             passwords = []
-        passwords = [None] + passwords + self.passwords
+        passwords = passwords + self.passwords
         combined_attachment_matches = []
         if isinstance(attachments, dict):
             attachments = [attachments]
@@ -490,22 +488,20 @@ class MailScanner(object):
         verdict = None
         multi_auth_headers = self.allow_multiple_authentication_results
         use_og_auth_results = self.use_authentication_results_original
-        trusted_domain = from_trusted_domain(
-            parsed_email, self.trusted_domains,
+        yara_safe_optional_domain = from_trusted_domain(
+            parsed_email, self.yara_safe_optional_domains,
             allow_multiple_authentication_results=multi_auth_headers,
             use_authentication_results_original=use_og_auth_results,
-            include_sld=self.include_sld_in_auth_check
         )
-        trusted_domain_yara_safe_required = from_trusted_domain(
-            parsed_email, self.trusted_domains_yara_safe_required,
+        authenticated_domain = from_trusted_domain(
+            parsed_email, [parsed_email["from"]["domain"]],
             allow_multiple_authentication_results=multi_auth_headers,
             use_authentication_results_original=use_og_auth_results,
-            include_sld=self.include_sld_in_auth_check
         )
-        auth_optional = False
         categories = []
         has_attachment = len(attachments) > 0
         for match in matches:
+            auth_optional = False
             if "no_attachments" in match["meta"]:
                 if match["meta"]["no_attachments"] and has_attachment:
                     continue
@@ -513,33 +509,29 @@ class MailScanner(object):
                 if match["meta"]["no_attachment"] and has_attachment:
                     continue
             if "from_domain" in match["meta"]:
-                sld = parsed_email["from"]["sld"]
-                if sld != get_sld(match["meta"]["from_domain"]):
+                domain = parsed_email["from"]["domain"]
+                if domain != match["meta"]["from_domain"]:
                     continue
             if "category" in match["meta"]:
-                categories.append(match["meta"]["category"])
                 if match["meta"]["category"] == "safe":
-                    if "auth_optional" in match["meta"] and not auth_optional:
+                    if "auth_optional" in match["meta"]:
                         auth_optional = match["meta"]["auth_optional"]
+                    authenticated = any([yara_safe_optional_domain,
+                                         auth_optional, authenticated_domain])
+                    if not authenticated:
+                        categories.append("yara_safe_auth_fail")
+                        continue
+
+                categories.append(match["meta"]["category"])
+
         categories = _deduplicate_list(categories)
         if len(categories) == 1:
             verdict = categories[0]
         elif len(categories) > 1:
             verdict = "ambiguous"
-        authenticated = any([trusted_domain, trusted_domain_yara_safe_required,
-                             auth_optional])
-        if verdict == "safe" and not authenticated:
-            verdict = "yara_safe_auth_fail"
-        if verdict != "safe" and trusted_domain_yara_safe_required:
-            verdict = "auth_pass_not_yara_safe"
-        if verdict is None and authenticated:
-            verdict = "safe"
 
-        safe_required = trusted_domain_yara_safe_required
         return dict(matches=matches, categories=categories,
                     msg_from_domain=msg_from_domain,
-                    trusted_domain=trusted_domain,
-                    trusted_domain_yara_safe_required=safe_required,
-                    auth_optional=auth_optional,
+                    msg_from_domain_authenticated=authenticated_domain,
                     has_attachment=has_attachment,
                     verdict=verdict)