+++ CHANGELOG.md
@@ -11,14 +11,13 @@ This release is a major rewrite that includes changes breaking existing use
     - `auth_pass_not_yara_safe` verdict removed
     - `trusted_domain`, `trusted_domain_yara_safe_required`, and `auth_optional` removed from results
   - The `auth_optional`rule `meta` value only applies to that rule
-  - Including the second-level domain (SLD) in authentication checks is now set by the `include_sld` rule `meta` value
   - Trusted domains are now called YARA optional domains
 - API changes
-  - `trusted_domains` renamed to `yara_optional_domains`
+  - `trusted_domains` renamed to `yara_safe_optional_domains`
   - `trusted_domains_yara_safe_required` parameter removed
   - `include_sld_in_auth_check` parameter removed
 - CLI changes
-  - `--trusted-domains` renamed to `--yara-optional-domains`
+  - `--trusted-domains` renamed to `--yara-safe-optional-domains`
   - `--trusted-domains-yara` removed
   - `--sld` removed